Fixed some stuff moved some stuff
