+++ src/main/java/matgm50/mankini/Mankini.java
@@ -1,5 +1,6 @@
 package matgm50.mankini;
 
+import matgm50.mankini.init.ModCreativeTab;
 import matgm50.mankini.init.ModConfigGen;
 import matgm50.mankini.init.ModEntities;
 import matgm50.mankini.init.ModItems;
@@ -8,7 +9,6 @@ import matgm50.mankini.proxy.CommonProxy;
 import matgm50.mankini.util.BatHandler;
 import matgm50.mankini.util.BatMankiniJump;
 import matgm50.mankini.util.DropHandler;
-import matgm50.mankini.util.TabMankini;
 import matgm50.mankini.util.TickHandler;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraftforge.common.MinecraftForge;
@@ -30,7 +30,7 @@ public class Mankini {
     @SidedProxy(clientSide = ModLib.CLIENTPROXY, serverSide = ModLib.COMMONPROXY)
 	public static CommonProxy proxy;
 
-    public static CreativeTabs tabMankini = new TabMankini(ModLib.MOD_ID);
+    public static CreativeTabs tabMankini = new ModCreativeTab(ModLib.MOD_ID);
     
     @EventHandler
     public void preInit(FMLPreInitializationEvent event) {
