+++ src/main/java/matgm50/mankini/init/ModRecipes.java
@@ -8,7 +8,7 @@ public class ModRecipes {
         //        '1', new ItemStack(Items.DYE, 1, 13), '2', new ItemStack(Items.DYE, 1, 5), '3', new ItemStack(Items.DYE, 1, 4), '4', new ItemStack(Items.DYE, 1, 2),
         //        '5', new ItemStack(Items.DYE, 1, 11), '6', new ItemStack(Items.DYE, 1, 14), '7', new ItemStack(Items.DYE, 1, 1));
 
-      //  GameRegistry.addShapedRecipe(new ItemStack(ModItems.itemAethericMankini), "X X", "DOD", " D ", 'X',  new ItemStack(Items.nether_star), 'D', new ItemStack(Blocks.diamond_block),
+    	//  GameRegistry.addShapedRecipe(new ItemStack(ModItems.itemAethericMankini), "X X", "DOD", " D ", 'X',  new ItemStack(Items.nether_star), 'D', new ItemStack(Blocks.diamond_block),
         //        'O', new ItemStack(ModItems.itemKawaiiMankini));
         
     }
